Week Three's Forecasts!
Now that we have a good system, and everybody was available, I'm internally fairly confident that this is close to the best we can do without either more manpower, more commitment from the existing people, or important additional learnings! 

Though I hope to improve on some of those factors (especially additional learnings)!